perf(tableOfContents): ⚡️ load the table of contents script only on pages that have one
The table of contents scroll spy and its section observer shipped in
the boot bundle on every page view. They now live in a separate
skins.citizen.toc module that only loads on pages that render a ToC,
cutting the boot bundle from 60.0 KB to 38.9 KB of minified JS
(14.5 KB to 9.8 KB gzipped) everywhere else — the main page, special
pages, categories, histories, diffs, and heading-less articles.

The trade-off is the module's startup manifest entry, measured at
35 bytes raw (11 bytes gzipped) on every page view; the split pays
for itself once ~0.3% of page views lack a ToC.

The server queues the module on ToC pages, so it rides the initial
batched module request with no extra round trip. A client-side element
check in setupObservers covers page HTML cached before the server-side
condition existed; mw.loader dedupes the two requests into one load.
If the module fails to load, the failure is logged and the
server-rendered ToC stays navigable without the scroll spy.

During the brief startup-manifest cache window after deploying, pages
rendered with the new HTML may log an unknown-module notice while the
old boot bundle (which still contains the scroll spy) is served; the
ToC keeps working through that bundle until the manifest refreshes.
