+++ tests/vitest/skins.citizen.scripts/setupObservers.test.js
@@ -4,102 +4,43 @@
 const setupObservers = require( '../../../resources/skins.citizen.scripts/setupObservers.js' );
 const mw = require( '../mocks/mw.js' );
 
-/**
- * Minimal ToC + article fixture: one top-level section whose heading
- * carries the id the location hash points at.
- *
- * @return {{ tocItem: HTMLElement }}
- */
-function createFixture() {
-	document.body.innerHTML = `
-		<div id="citizen-toc">
-			<div class="citizen-toc-indicator"></div>
-			<ul id="mw-panel-toc-list">
-				<li id="toc-s1" class="citizen-toc-list-item citizen-toc-level-1">
-					<a class="citizen-toc-link" href="#s1">Section 1</a>
-					<button class="citizen-toc-toggle"></button>
-				</li>
-				<li id="toc-s2" class="citizen-toc-list-item citizen-toc-level-1">
-					<a class="citizen-toc-link" href="#s2">Section 2</a>
-					<button class="citizen-toc-toggle"></button>
-				</li>
-			</ul>
-		</div>
-		<div id="bodyContent">
-			<div class="mw-parser-output">
-				<div class="mw-heading"><h2 id="s1">Section 1</h2></div>
-				<div class="mw-heading"><h2 id="s2">Section 2</h2></div>
-			</div>
-		</div>
-		<div id="citizen-page-header-sticky-sentinel"></div>
-	`;
+function createMockWindow() {
 	return {
-		tocItem: document.getElementById( 'toc-s1' ),
-		tocItem2: document.getElementById( 'toc-s2' ),
-		heading2: document.querySelectorAll( '.mw-heading' )[ 1 ]
+		addEventListener: vi.fn(),
+		removeEventListener: vi.fn(),
+		getComputedStyle: vi.fn( () => ( {
+			getPropertyValue: vi.fn( () => 'block' )
+		} ) )
 	};
 }
 
-describe( 'setupObservers', () => {
-	let rafQueue;
-	let idleCallbacks;
-	let win;
-
-	function flushFrame() {
-		rafQueue.splice( 0 ).forEach( ( cb ) => cb() );
-	}
-
-	function createMockWindow() {
-		return {
-			addEventListener: vi.fn(),
-			removeEventListener: vi.fn(),
-			location: { hash: '#s1' },
-			innerHeight: 800,
-			scrollY: 0,
-			matchMedia: vi.fn( () => ( {
-				// Desktop breakpoint matches so the scroll spy is active
-				matches: true,
-				addEventListener: vi.fn(),
-				removeEventListener: vi.fn()
-			} ) ),
-			getComputedStyle: vi.fn( () => ( {
-				getPropertyValue: vi.fn( () => '75' )
-			} ) ),
-			requestAnimationFrame: vi.fn( ( cb ) => cb() )
-		};
-	}
-
-	beforeEach( () => {
-		rafQueue = [];
-		// deferUntilFrame uses the global requestAnimationFrame
-		vi.stubGlobal( 'requestAnimationFrame', ( cb ) => {
-			rafQueue.push( cb );
-			return rafQueue.length;
-		} );
-
-		idleCallbacks = [];
-		mw.requestIdleCallback = vi.fn( ( cb ) => {
-			idleCallbacks.push( cb );
-		} );
+function createMockIntersectionObserver() {
+	return vi.fn( function () {
+		this.observe = vi.fn();
+		this.unobserve = vi.fn();
+		this.disconnect = vi.fn();
 	} );
+}
 
+describe( 'setupObservers', () => {
 	afterEach( () => {
-		[ 've.activationStart', 've.deactivationComplete', 'wikipage.tableOfContents' ]
+		[ 've.activationStart', 've.deactivationComplete' ]
 			.forEach( ( name ) => mw.hook( name )._reset() );
-		vi.unstubAllGlobals();
 		vi.clearAllMocks();
 		document.body.innerHTML = '';
 	} );
 
-	it( 'should defer initial section activation until idle plus one frame', () => {
-		const { tocItem } = createFixture();
-		win = createMockWindow();
-		mw.util.getTargetFromFragment = vi.fn( () => tocItem );
-		const MockIntersectionObserver = vi.fn( function () {
-			this.observe = vi.fn();
-			this.unobserve = vi.fn();
-			this.disconnect = vi.fn();
-		} );
+	it( 'should lazy-load the table of contents module when the page has a ToC', async () => {
+		document.body.innerHTML = `
+			<div id="citizen-toc"></div>
+			<div id="bodyContent"></div>
+			<div id="citizen-page-header-sticky-sentinel"></div>
+		`;
+		const win = createMockWindow();
+		const MockIntersectionObserver = createMockIntersectionObserver();
+		const tocInit = vi.fn();
+		const moduleRequire = vi.fn( () => ( { init: tocInit } ) );
+		mw.loader.using.mockImplementation( () => Promise.resolve( moduleRequire ) );
 
 		setupObservers.init( {
 			document,
@@ -108,31 +49,68 @@ describe( 'setupObservers', () => {
 			IntersectionObserver: MockIntersectionObserver
 		} );
 
-		// Boot idle tasks dirty document-level state; reading TOC geometry
-		// before they are painted forces a full-page reflow. Nothing may
-		// activate synchronously.
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--expanded' ) ).toBe( false );
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( false );
-		expect( idleCallbacks.length ).toBeGreaterThan( 0 );
+		expect( mw.loader.using ).toHaveBeenCalledWith( 'skins.citizen.toc' );
 
-		idleCallbacks.forEach( ( cb ) => cb() );
+		await Promise.resolve();
 
-		// Still waiting for the post-idle frame
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( false );
-
-		flushFrame();
-
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--expanded' ) ).toBe( true );
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( true );
+		expect( moduleRequire ).toHaveBeenCalledWith( 'skins.citizen.toc' );
+		expect( tocInit ).toHaveBeenCalledWith( {
+			document,
+			window: win,
+			mw,
+			IntersectionObserver: MockIntersectionObserver
+		} );
 	} );
 
-	it( 'should defer the hash-less initial intersection calculation to idle plus one frame', () => {
-		createFixture();
-		win = createMockWindow();
-		win.location.hash = '';
+	it( 'should log instead of rejecting when the module fails to load', async () => {
+		document.body.innerHTML = `
+			<div id="citizen-toc"></div>
+			<div id="bodyContent"></div>
+			<div id="citizen-page-header-sticky-sentinel"></div>
+		`;
+		const win = createMockWindow();
+		const loadError = new Error( 'Unknown module' );
+		mw.loader.using.mockImplementation( () => Promise.reject( loadError ) );
+		mw.log = { warn: vi.fn() };
+
+		setupObservers.init( {
+			document,
+			window: win,
+			mw,
+			IntersectionObserver: createMockIntersectionObserver()
+		} );
+		await Promise.resolve();
+
+		expect( mw.log.warn ).toHaveBeenCalledWith(
+			'Failed to load skins.citizen.toc', loadError
+		);
+	} );
+
+	it( 'should not load the table of contents module when the page has no ToC', () => {
+		document.body.innerHTML = `
+			<div id="bodyContent"></div>
+			<div id="citizen-page-header-sticky-sentinel"></div>
+		`;
+		const win = createMockWindow();
+
+		setupObservers.init( {
+			document,
+			window: win,
+			mw,
+			IntersectionObserver: createMockIntersectionObserver()
+		} );
+
+		expect( mw.loader.using ).not.toHaveBeenCalled();
+	} );
+
+	it( 'should still observe the sticky header sentinel without a ToC', () => {
+		document.body.innerHTML = `
+			<div id="bodyContent"></div>
+			<div id="citizen-page-header-sticky-sentinel"></div>
+		`;
+		const win = createMockWindow();
 		const ioInstances = [];
-		const MockIntersectionObserver = vi.fn( function ( cb ) {
-			this.cb = cb;
+		const MockIntersectionObserver = vi.fn( function () {
 			this.observe = vi.fn();
 			this.unobserve = vi.fn();
 			this.disconnect = vi.fn();
@@ -146,80 +124,10 @@ describe( 'setupObservers', () => {
 			IntersectionObserver: MockIntersectionObserver
 		} );
 
-		// The section observer (first constructed) only observes elements when
-		// calcIntersection runs — which must wait for idle plus one frame
-		const sectionSpyIo = ioInstances[ 0 ];
-		expect( sectionSpyIo.observe ).not.toHaveBeenCalled();
-
-		idleCallbacks.forEach( ( cb ) => cb() );
-		expect( sectionSpyIo.observe ).not.toHaveBeenCalled();
-
-		flushFrame();
-
-		expect( sectionSpyIo.observe ).toHaveBeenCalled();
-	} );
-
-	it( 'should skip the deferred initial activation when the scroll spy has already fired', () => {
-		const { tocItem, tocItem2, heading2 } = createFixture();
-		win = createMockWindow();
-		// Page loaded with a hash pointing at section 1
-		win.location.hash = '#s1';
-		mw.util.getTargetFromFragment = vi.fn( () => tocItem );
-		const ioInstances = [];
-		const MockIntersectionObserver = vi.fn( function ( cb ) {
-			this.cb = cb;
-			this.observe = vi.fn();
-			this.unobserve = vi.fn();
-			this.disconnect = vi.fn();
-			ioInstances.push( this );
-		} );
-
-		setupObservers.init( {
-			document,
-			window: win,
-			mw,
-			IntersectionObserver: MockIntersectionObserver
-		} );
-
-		// Before idle fires, the user scrolls and the spy activates section 2
-		ioInstances[ 0 ].cb( [
-			{ target: heading2, boundingClientRect: { top: 10 } }
-		] );
-		expect( tocItem2.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( true );
-
-		idleCallbacks.forEach( ( cb ) => cb() );
-		flushFrame();
-
-		// The stale boot-time hash target must not clobber the live spy state
-		expect( tocItem2.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( true );
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( false );
-	} );
-
-	it( 'should not activate at idle on narrow viewports where the spy is paused', () => {
-		const { tocItem } = createFixture();
-		win = createMockWindow();
-		// Below the desktop breakpoint the ToC popover is collapsed
-		win.matchMedia = vi.fn( () => ( {
-			matches: false,
-			addEventListener: vi.fn(),
-			removeEventListener: vi.fn()
-		} ) );
-		mw.util.getTargetFromFragment = vi.fn( () => tocItem );
-		const MockIntersectionObserver = vi.fn( function () {
-			this.observe = vi.fn();
-			this.unobserve = vi.fn();
-			this.disconnect = vi.fn();
-		} );
-
-		setupObservers.init( {
-			document,
-			window: win,
-			mw,
-			IntersectionObserver: MockIntersectionObserver
-		} );
-		idleCallbacks.forEach( ( cb ) => cb() );
-		flushFrame();
-
-		expect( tocItem.classList.contains( 'citizen-toc-list-item--active' ) ).toBe( false );
+		const sentinel = document.getElementById( 'citizen-page-header-sticky-sentinel' );
+		const observed = ioInstances.some(
+			( io ) => io.observe.mock.calls.some( ( args ) => args[ 0 ] === sentinel )
+		);
+		expect( observed ).toBe( true );
 	} );
 } );
